requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cors-headers==3.3.0
django-extensions==2.2.9
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
